Move GuildMemberUpdateNicknameEvent
Extend GenericGuildMemberUpdateEvent and modify constructor to support it
